+++ cmdeploy/src/cmdeploy/basedeploy.py
@@ -51,7 +51,7 @@ def get_resource(arg, pkg=__package__):
     return importlib.resources.files(pkg).joinpath(arg)
 
 
-def configure_remote_units(deployer, mail_domain, units) -> None:
+def configure_remote_units(deployer, mail_domain, units, **kwargs) -> None:
     remote_base_dir = "/usr/local/lib/chatmaild"
     remote_venv_dir = f"{remote_base_dir}/venv"
     remote_chatmail_inipath = f"{remote_base_dir}/chatmail.ini"
@@ -63,6 +63,7 @@ def configure_remote_units(deployer, mail_domain, units) -> None:
             config_path=remote_chatmail_inipath,
             remote_venv_dir=remote_venv_dir,
             mail_domain=mail_domain,
+            **kwargs,
         )
 
         basename = fn if "." in fn else f"{fn}.service"
@@ -226,7 +227,7 @@ class Deployer:
         res = files.directory(name=name, path=path, present=False, **kwargs)
         return self._update_restart_signals(path, res)
 
-    def download_executable(self, url, dest, sha256sum, extract=None):
+    def download_executable(self, url, dest, sha256sum, extract=None, mode="755"):
         existing = host.get_fact(Sha256File, dest)
         if existing == sha256sum:
             return
@@ -243,7 +244,7 @@ class Deployer:
                 f"({dl_cmd}"
                 f" && echo '{sha256sum}  {tmp}' | sha256sum -c"
                 f" && mv {tmp} {dest})",
-                f"chmod 755 {dest}",
+                f"chmod {mode} {dest}",
             ],
         )
         self.need_restart = True